backend/requirements.txt
Flask==2.1.0 
Flask-SQLAlchemy==2.5.1 
python-dotenv==0.19.2 
Flask-Cors==4.0.0 
Flask-JWT-Extended==4.5.2 
